Introduce LangLinkHandlerFactory
LangLinkHandlerFactory is useful to make possible to inject dependencies into LangLinkHandler based on the page rendered.
For example, it allows to inject a UsageAccumulator to LangLinkHandler that is useful if ones want to display site links
not from the current item sitelinks but from a statement or an other item.

Bug: T128173
Change-Id: I3bee7d7c25a8cd51357efbb38227f7b50a9d72da