Introduce SiteLinksForDisplayLookup
This class is used to provide the list of sitelinks to display
with a given Wiki page. It is useful to have a single source
for the list of the site links to display, list that might
be expended using a hook.

This hook addition would be useful to e.g. add sitelinks from a connected items. This use case is required to use in Wikisource not only the site links of the "edition" Wikidata items but also the site links of the connected "work" item.

Bug: T128173
Change-Id: I65543c925c4cdf2acf235d2cfcab75bc94a384e2