Move PREFETCHING_TERM_LOOKUP definition to Lib
This was moved from SingleEntitySourceServices to Repo entitytypes in
e07cc1872caccc35ed93d0df4b71f9a19d498149 which removed clients access to
the service.

This didnt fail in a very visible way due to getPrefetchingTermLookup in
SingleEntitySourceServices falling back to a NullPrefetchingTermLookup.

This patch moves the logic needed into the Lib wiring and the few needed
services and data points into SingleEntitySourceServices.

This should fix the UBN on master but there should still be a few
followups to this:
 - Determine what to do with NullPrefetchingTermLookup and perhaps
 convert it to throw exceptions instead to indicate bad wiring.
 - An integration test for the case identified by the UBN will be added
 https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/Wikibase/+/594929/
 - Cleanup some of services added to SingleEntitySourceServices, it will
 likely make sense to put all of the terms storage services here, but
 perhaps via a factory dedicated for terms storage, similar to the
 TermStoreWriterFactory that will be introduced in
 21612ee355e725d05671b5c2afe140ae32a6d710 but for all access to the
 store.

This adds 2 already deprecated methods to SingleEntitySourceServices,
both of which will be removed once wb_terms is finally removed from
Wikbase code. I could make this nicer now, but it seems like a poor
investment considering the code is about to be removed anyway.

Bug: T252079
Change-Id: I9c5d70e01a647dd4f1aa395a418061e4d74b8a32