bridge: selenium: re-enable "go back from a save error" test
Hypothesis: the save button is clicked while the (CSS/JS) resize
animation is in progress and it effectively moves out from underneath
the mouse which results in its click to not have full effect (key up is
happening outside of the element, as suggested by the white outline in
the screenshot [0]).

Maybe one could waitUntil getWindowSize reports the desired value but
I'm not sure this would account for the animation and waiting for that
seems to require very deep understanding of what its internals are, if
it is possible at all.

[0]: https://phabricator.wikimedia.org/F31812692

Bug: T251538
Change-Id: Ifa52daf07f92aec1bd1a242421ff3d803b20a393