bridge: handle saving errors in root action
Committing SAVING_FAILED unconditionally in saveBridge() leads to noise
both in the error tracking and in the getter for assertUserFailedError
and also probably future similar getters. Doing the entire error
handling in saveBridge instead of splitting it with the entity module
action should make it more obvious what is going on.

Change-Id: Idc762479ae612866770f1bd0e1e77c1348b19530