bridge: selenium: add missing waitForDisplayed
As far as I can tell, this is the only test that launches the bridge,
then requires the bridge to be done loading, but doesn’t wait for the
bridge or error to be shown in between. (Most other tests already have a
similar waitForDisplayed; a few, such as canCancel, only need the app to
be ready and don’t need to wait for loading to be done.)

This arguably makes the immediately following assertion rather
redundant, but it mirrors other tests (e. g. “shows the current
targetValue”), so in my opinion this can be improved later.

Bug: T252449
Change-Id: Idbc70565ea0b0716d2ec56169c7b115e1da621d7