bridge: selenium: abstract way to set window size
Add a convenient way for tests to demand their desired window size -
"mobile" or not.

This will check if the desired setting is already in place and only
actually change the size if a switch in mode would be resulting.

It also ensures that possible animations that stem from the resizing are
awaited to save us from surprises in the tests.

Change-Id: I0e4abc4612dd8ea6affd27c290e6e0d794b458d8