bridge: add login warning message
Define a new message for the warning shown to users when we send them to
the login page and need them to close the tab afterwards, and register
it as a valid login warning message using the appropriate hook. The
message isn’t used yet, but I figured this was enough of a change to
warrant its own commit.

For the hook handler, we already implement the hook’s new interface,
anticipating that the modern MediaWiki hook container (see T240307) will
soon be ready. For the moment, though, the HookHandlers for that system
cannot yet be defined in extension.json (see T252686), so we still
include a static entry point and register that as the hook handler. We
can hopefully migrate this to the newest system soon, though.

Bug: T251395
Change-Id: I0c92d4074493d67a5615e8f737b96084c26433c4