WikibaseValueFormatterBuilders: inject specialized services
This injects four specialized services into
WikibaseValueFormatterBuilders instead of creating them inside of the
class. For Wikibase repo this means that the dispatching variants of
the services can be injected, which in turn unlocks the expected
behavior for Federated Properties.

For Wikibase client, the title lookup based implementations are
injected directly.

Bug: T251776
Change-Id: I9c613f53ea4f57b89d482a92634c5b7feeed3a56