Add populateSitesTable.php back to lib/autoload.php
Change If78aa50dad (commit 66911ea294) removed the lib/autoload.php
entry for this file, but it cannot actually be loaded via PSR-4: we
don’t register lib/maintenance/ as the autoload directory for the
Wikibase\Lib\Maintenance\ namespace (so the autoloader would instead
look in lib/includes/Maintenance/), and even if we did, the filename
doesn’t match the case of the class (class name begins with uppercase P,
file name begins with lowercase p).

We cannot fix this via AutoloadClasses in extension-lib-wip.json
(build/generateAutoload.php throws an exception on AutoloadClasses, and
we still need that script for a little while), so for now, revert back
to an autoload.php file to ensure that PopulateSitesTable can still be
autoloaded. Hopefully we can fix this eventually (possibly with T99268).

This partially reverts commit 66911ea294fc5bddebc6c62b7f9cd056b5a1244d.

Bug: T172368
Change-Id: Id7b52b31f8f4c558ea1945b31f3791abc5813b44