bridge: Move error formatting into tracker
The store should not need to know about the formatting conventions of
the tracker service.

Also, this reduces the likelihood of errors, because now all calls to
`errorTypeFormatter` are in one file instead of potentially all over the
code base.

It also adds a test case for
`DataBridgeTrackerService::trackSavingUnknownError`, which was
apparently missed when the feature was implemented.

Bug: T252683
Change-Id: I828310ae19c9ec8b376c85237e3938b6e9d3a202