bridge: update to latest @wmde/wikibase-vuejs-components
Update to latest version of @wmde/wikibase-vuejs-components. This takes
us from v0.1.7 to v0.1.9 - effective changes:
* Improve[d] keyboard accessibility [of `RadioInput`]
* Update wikimedia-ui-base, changing `.wb-ui-radio-input label` color
(#222 => #202122; I can't see the difference)
* Make RadioInput work in IE11

Our e2e tests simulate using the application (just short of browser
tests). So far they dispatched an 'input' event on the radio input to
execise the app. This, of course, does not work anymore because
RadioInput does not listen to 'input' events any more. Instead of merely
changing this to a 'change' event, we simulate a mouse event (as if a
real user clicked) and let the browser figure out which element events
result from that.

Bug: T249343
Change-Id: Icbcdccb9a2a14bca6e123ed787b9e1fe88c9c940