FP: ApiEntityLookup fetch and log when not already prefetched
Throwing an exception here broke things in a few places and we will have
to look into exactly where this prefetching happens in order to make all
of this a bit nicer.

We will likely want to continue logging (perhaps also injecting the
logger though) so that we can cleanup missed use cases.

It would be ideal to one day prefetch everything, however perhaps that is
unrealistic.

Bug: T253125
Change-Id: Iac15509585edd38fedce569b54bf8e3643fcb597