Use EntityUrlLookup to set hrefs in HtmlPageLinkRendererEnd
HtmlPageLinkRendererEnd is used because HtmlPageLinkRendererBegin does
not allow overriding the href attribute.

To avoid having to repeatedly decide whether a link should be formatted
as an entity link, the HtmlPageLinkRendererBeginHookHandler now sets a
custom attribute to links that get formatted as entity links. This
attribute is picked up by the new HtmlPageLinkRendererEndHookHandler to
override the href attribute which is necessary to correctly link to
federated properties.

Note that overriding an entity link's href attribute may not be
necessary if in the future non-local entities ids have interwiki
prefixes.

Bug: T252411
Change-Id: I53acb57fa77b360892b865e641d14910f1789c78