bridge: selenium: extract WarningAnonymousEdit page object
This reduces the size of dataBridge.page.js and also, I think, improves
readability, by getting rid of some overly large function/getter names.

Note that other page objects (for errors) import DataBridgePage itself
and make their root relative to that. However, since DataBridgePage
imports WarningAnonymousEdit (to auto-dismiss the warning), doing the
same in WarningAnonymousEdit would lead to a dependency cycle. Node has
limited support for those, but not if you overwrite the whole exports
object, as we do. Copy+pasting the root selector is a simple and
straightforward way to resolve this cycle, and seems acceptable to me.
(Perhaps at some point DataBridgePage will be split up enough that we
can import the root selector without a cycle, but for now this will do.)

Bug: T253208
Change-Id: Icfb09c3d8c6f2bf96f91d39450320b0a0bfe101e