bridge: split app into modern and legacy build
vue-cli can build projects with a --modern flag [1], which produces two
JS bundles: one for modern browsers (defined as supporting JavaScript
modules [2]), and one for legacy browsers. vue-cli’s index.html
distinguishes between these cases using <script> tags with type="module"
or nomodule attributes, which doesn’t directly work for us, but we can
build something similar – that’s what load-app.js in this change does.
The modern and legacy bundles are exposed as separate ResourceLoader
modules, and the main .app module only contains load-app.js, which
dynamically loads the modern or legacy bundle. (This means that the app
module export is now a *Promise* resolving to the real app module, not
the app module immediately, so the init module has to `await` the
require()d value.)

In principle, we could probably do something similar for the init
module, but let’s start with the app module first, which is larger and
offers the higher benefit. Comparison for the app module:

without gzip   legacy bundle   187.41 KiB
               modern bundle   155.82 KiB   Δ-17%
   with gzip   legacy bundle    47.14 KiB
               modern bundle    41.38 KiB   Δ-12%

(“gzip” refers to `gzip` on the command line without options, i. e. the
default compression level 6. I don’t know how it’s configured in prod.)

Note that we still ship styles and messages with the main .app module,
since they don’t differ between the modern and legacy modules and their
effects are global anyways.

(vue-cli adds a tiny JS snippet to the HTML to ensure that Safari 10.1,
which supports modules but not the nomodule attribute, doesn’t load both
bundles; we don’t have anything like that, Safari 10.1 will just be
treated as a legacy client. Recent Safari versions are not affected.)

[1]: https://cli.vuejs.org/guide/browser-compatibility.html#modern-mode
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules

Bug: T253846
Change-Id: Ifbc5a6d64632add537f9191f91407cd1dabd0d84