bridge: Better typing for module boundary signatures
This adds types to signatures of methods that are on the boundary of
modules. This satisfies the explicit-module-boundary-types rule that was
added in 2.17.0 version of typescript-eslint and becomes part of the
recommended set of rules in version 3.0.0.

This rule, like similar rules, is deactivated for .js file as it would
produce false positives there.

Change-Id: Ibe0a4f49eb69a1981367be6598c682d0fdc4c60a