bridge: ignore if reference rendering fails
So far the application never left the loading state if rendering the
references failed. This is most certainly not the desired behavior.

The question was how we deal with it (show error or somewhat gracefully
degradate) - product decided to ignore the error and proceed with bridge
as if there are not references.

To gain some insights if/how often this happens, tracking is added.

Bug: T253857
Change-Id: I82f388de59b1517e54c8cf99b09738137480d055