Use ThrowingEntityTermStoreWriter when writers shouldn't be called
Item and Property handler classes need access to their TermStoreWriters,
as the handlers deal with a mixture of reading and writing things.
The refactoring in 710be17603451467ee75605e1e4210e342690a5d as part of
T252019 did not take into account the fact that writers would need to be
created in situations where no writing was expected to happen.

A few alternative solutions were considered such as:
 - Pass in a set of factories to the handlers and let them do the
 construction. This is complex however as only 1 set of writers is
 created using a actual factory currently. so a mix of factories and
 callbacks would be needed?
 - Just pass in callbacks to get the writers? Sounds less testable and
 quite evil

The final solution is to create the handlers with a throwing writer when
writing should not happen (non local entity source).
This is throwing so that if something gets messed up it will explode
rather than silently doing unexpected unknown things.

This area of code will get a little nicer when we finish removing the
legacy wb_terms stuff, and at that point passing in a factory might be
nicer than wiring up a throwable service.

Keen for this to be ready for backporting and consider the best testing
strategy after. This could be a general approach to testing WikibaseRepo
services with non local entity sources in a more complete / thorough
way.

Bug: T253804
Change-Id: I1b26518e87be6b92b758638892c67b5a1e9cdcfe