bridge: use TypeScript readonly arrays
TypeScript 3.4 added `readonly X[]` syntax as syntactic sugar for the
already available `ReadonlyArray<X>` type. This change uses that to turn
many regular arrays into readonly arrays. I hope that this will improve
readability, making it clearer which arrays are supposed to change over
time and which aren’t, and maybe prevent some accidental modifications.

As a general rule, I’ve mainly added the readonly modifier to function
parameters, to object/interface members, and (except in tests, where I
don’t think it’s worth it) to local variables, but not to function
return types. My interpretation is that a function declaring a parameter
readonly is like saying to the caller “I promise that I won’t modify
this array you give me”, whereas declaring the return type readonly is
more like “you must not modify this array I’m giving you” (e. g. because
something else also holds a reference to the array); to me this feels
similar to the `const` modifier in C, or to ownership vs. borrowing in
Rust. Declaring a parameter as readonly is therefore good and useful
(and has no effect on how the function can be called), whereas declaring
the return type as readonly places additional restrictions on callers.
That said, in some cases it was necessary to make a returned array
readonly because it’s directly derived from a readonly input array.

Bug: T253211
Change-Id: Ia033044f886a8ac10243fc0719a03e44cd7d6a45