bridge: don't build dev entrypoint code as library
Iaed28cbddb212e404dca02844750db2ae622ca77 changed the target format for
all entrypoints to library/commonjs - but this did not account for the
fact that the dev entrypoint can not deal with that.

It is unclear to me, if this is the case, why there are no built-in
safeguards to override this for the dev server - but a quick websearch
did not yield any insights.

It looks tempting to put the `--target lib --formats commonjs` back into
the build script of the package.json but apparently this does not work,
as the cli command then insists on the relevant entrypoint to be
specified - but we have two and want to build both.

Bug: T253202
Change-Id: I533e6a8f0322de142c37dfee143de9478106ca34