bridge: tell ResourceLoader not to minify our resources
By default, ResourceLoader minifies all the sources it serves (except in
debug mode). While there isn’t much benefit in doing this for our
resources, which were already minified at build time (using a more
efficient minifier – ResourceLoader’s has to be fast enough to run
during a web request), it usually doesn’t hurt either – except that we
recently started using vue-cli’s modern mode. This means that we’re now
serving some ES6 sources (e. g. using => syntax), even though
ResourceLoader’s minifier only supports ES5 syntax, which can cause
resources to be broken by the minifier. A /*@nomin*/ comment can be used
to instruct ResourceLoader to not minify a resource; use the webpack
BannerPlugin to add this to all of our resources.

Note that the default configuration of the plugin would result in a
comment like this:

    /*! /*@nomin* / */

But we need the source code to contain *exactly* /*@nomin*/
(ResourceLoader looks for this marker using a simple strpos()). On the
other hand, block comments that don’t begin with /*! are stripped from
the build, at a stage after the banner plugin runs, so we can’t just use
/*@nomin*/ itself either. Instead, use the string /*!/*@nomin*/ – the
/*! tells webpack not to remove this comment, /* is part of the
ResourceLoader marker (but not actually the beginning of a comment, but
rather the content of the already started /*! comment), and then */ is
both part of the marker and ends the real /*! comment.

Bug: T254261
Change-Id: Ia9605cc36ce91c0493f75eb5749dee38c5526475