Don't use EntityInfo in ViewFactory
This will allow WikibaseLexeme to now pass null into ViewFactory
newStatementSectionsView and newStatementGroupListView as the type hints
have been removed, and that is needed to do the migration / removal
dance between the two extensions

One this change is merged all extensions that have view factory
callbacks will also be able to remove their typehints and add default
values of null for the entity info objects preparing for their removal.

The view factory callbacks now have a default value of null, which can
be safely passed down into the views.
Although callers currently still get and pass down EntityInfo projects,
which is also okay everywhere, this will be removed in a followup.

Further removal will happen in a followup.

Bug: T254283
Change-Id: I3801b4d1f4a876a820c28f4541ec0802d886b7eb
Depends-On: I4615c32cefc2e1af685137cb96c37e3813bdce75