bridge, TR doc: take into account file removals
Ensure that data of file removals makes its way into the visualization
software so that "gaps" can be shown accordingly.

Adjusts size delta calculation to accomodate for this kind of event
(i.e. ignores both creations and removals as both would make for a 100%
delta which would make the user none the wiser)

Bug: T253204
Change-Id: Ic429daf98ac1773b5b0cb5bd99c137d907679955