Refactor LATEST_FROM_* to LookupCnstants
During work in T251878 the requirement to pass a lookupMode down to
RevisionBasedEntityLookup with the changes going all the way down to the
Store. During discussions the consensus was that Store should not have a
dependency to a constant in EntityRevisionLookup, hence the constants
should move to the Store interface.

After some more work re realized that these constants were needed outside of repo
and Store was currently in repo, thus the creation of LookupConstants which is now used.

Bug: T254443
Change-Id: I0fc795aae9c65bbd73cf0ff34a94b1d4642d43fd
Depends-On: I90f350c343e01260dd2411b1c3fd5856acbf63aa
Depends-On: I62f8c3be3af7382380a9fad45061fbd8c74347cf