doc: dist-size: run requests and I/O in parallel
The GitHub API doesn’t seem to have a limit on parallel requests (only
on requests overall), so we can run the analysis for all files in
parallel. On my system, this speeds up the analysis by roughly a third.

(Also parallelize the file I/O at the end, though it probably doesn’t
make much of a difference.)

Bug: T253204
Change-Id: Iee977fa0ee2616e5230ff9054af2e831b0f9b469