Fix watching new entities on existing pages
If a user’s preferences said to watch newly created pages, but not
edited pages, Wikibase would previously watch a page any time an entity
was created on it, even if the page itself had already existed
previously. This is common, for instance, with WikibaseMediaInfo, where
a file may exist but not have any structured data yet.

I could not find a satisfactory way to test this within Wikibase (the
base Wikibase data model entity types do not allow for item or property
pages to exist without an entity in them), but I added an integration
test in WikibaseMediaInfo instead, see change I480f96f2ea.

The comment on isNew() is reduced because, as far as I can tell, it was
inaccurate (it’s possible for getLatestRevisionId() to return 0 even if
the Title already exists, as long as the entity does not exist yet), and
we probably want the actual behavior, not the described behavior.

The check at the end of getWatchDefault() is changed from isNew() to
getEntityId() !== null, because I think that is the condition that is
actually necessary in order to call entityStore->isWatching(). (This is
more obvious in the older version of the code before it was changed to
use EntityStore – compare change Ieb3be710c6, commit 58d27b73c6.)

Bug: T216369
Change-Id: Ib19df094e8d1ceeb15df9ff7ecb92d277390ccaa