WikibaseClient: Use SkinAfterPortlet hook to add languages action links
WikibaseClient uses two hooks to expose action links in the sidebar
languages section. These are SkinTemplateOutputPageBeforeExec and
BaseTemplateAfterPortlet hook.

The former does some redundant checks (see below) and calls the method
that will generate the links,it then pushes the result into skin template
data (if non null). The other hook retrieves from the template data and
append it to the portlet.

Both hooks are deprecated, and SkinAfterPortlet hook can do all the work

The redundant checks are:
1. Title existence,
2. Action type,
3. Wikibase enabled namespace check, and
3. Check for {{noexternallanglinks}} parser function.

In rewriting this I believe these are no longer needed. After all these
checks pass, the hook call RepoItemLinkGenerator::getLink() method where
these checks are repeated entirely. This need not be. This patch now uses
the return value from the link generator which is guaranted to be either
string (add or edit action html fragments) or null.

Wikibase no longer manipulates the skin 'language_urls' data to set it
to empty array if there are no languages. If a page passes all these
checks but has no languages link, RepoItemLinkGenerator::getLink() will
still return string (for 'edit' action). Appending the string to the
'lang' portlet is now enough to trigger skins to diplay the section title
and append the html string (previously it was not).

Bug: T252841
Depends-On: I438daa79d3d97e2518e6258c3213a805bd1f30e8
Change-Id: Iff2ae0a41813d79f8b0a327b4e1cf8d4d4a0f797