Move Store to right namespace
It’s used in WikibaseLexeme and WikibaseQualityConstraints, so add a
temporary class alias. And as in change I75135d6d60 (commit ba3a68c0da),
not just that alias but also the new name needs to be added to
autoload.php for AutoLoaderStructureTest to be happy.

One bogus usage of Store in PrefetchingWikiPageEntityMetaDataAccessor,
which made NoBadDependencyUsageTest fail now that Store can be seen to
be part of Repo, is removed. I assume it was auto-imported at some point
from the word “Store” in a comment near the end of the file.

Bug: T172368
Change-Id: I1340569f813271d6acd758e151e43af997c35d8a