Stop generating EntityInfo in ParserOutputGenerator
This patch stops generating EntityInfo objects when rendering pages and
also stops passing the EntityInfo objects down into view factory
services (in extensions).
The extensions were modified in previous patches so that all callbacks
no longer have a type hint and have a default value of null.

This should remove all of the functional uses of this EntityInfo stuff,
although the classes remain and will be removed in a followup to make
this easier to review.

This patch will likely speed up all page renders when deployed as this
prefetching of entity info step will no longer happen.
It will also reduce memcached used on wikidata.

Depends-On: I1edec8dfd35507acfc7a6ba8d95838d5f6255d41
Depends-On: I727df0153b11c20c17ca26179a619b2b5728fa05
Bug: T254283
Change-Id: I10b76212615a140aceeaf2200221e9d5dedfa3a0