Move some \Wikibase\Rdf classes to \Wikibase\Repo\Rdf
These are all the classes in that namespace which aren’t used in any
other extension. The other classes will be moved soon as well, but
separately.

One effect of removing these classes from autoload.php is that they will
now only be autoloadable once MediaWiki has processed the extension
registration for WikibaseRepository (which does not happen with the
wfLoadExtension() call yet – that only enqueues the extension). Mostly,
this isn’t significant, but there is one case where it matters:
WikibaseRepo.datatypes.php uses constants from PropertyRdfBuilder, and
that code is evaluated immediately in repo/Wikibase.php, not as part of
any callback, so PropertyRdfBuilder needs to be autoloadable at that
point already. We fix this by adding that class to $wgAutoloadClasses;
we could just as well copy the full AutoloadNamespaces from the
extension JSON file (it’s not much, after all), but I want to keep this
fix very limited, and avoid other uses of autoloaded classes before the
extension is fully registered.

Bug: T172368
Change-Id: I987d4377347f4a24a4aedfa4ab3512c20911a26d