Defer evaluating rdf-data-type DataTypeDefinitions field
This gets rid of the hack introduced in change I987d437734: by moving
the rdf-data-type into a callback, the PropertyRdfBuilder class no
longer has to be autoloaded as soon as WikibaseRepo.datatypes.php is
evaluated, but rather only when the callback is called, at which time
extension registration is done and the autoload namespaces are
registered. We continue to allow un-callback’ed rdf-data-type values,
just in case any are used in third-party extensions, though I’m not
aware of any uses outside of Wikibase (and found none with code search).

Change-Id: I756c1c0091ab2dd5791ddd7515890c6483dfc2d2