Remove the last autoload.php
Two maintenance scripts must be autoloadable for DatabaseSchemaUpdater,
and are added to AutoloadClasses; however, we anticipate that they can
be removed soon, once we remove support for the wb_terms table.

Six maintenance scripts need to be loaded in some way in order for their
unit tests to work; however, we want to avoid any other code depending
on those maintenance scripts, so we only load them in the test files
themselves, using require_once.

Bug: T172368
Bug: T255108
Change-Id: I0f52629775144949eb8140619dc1e2f03f6f97cc