Make nasty hack slightly less nasty
A handful of Lib tests currently depend on Repo classes; to make them
work on Client-only wikis, we register those classes for autoloading in
Lib. Do this in the JSON file using TestAutoloadClasses, rather than in
the PHP entry point. (This means that the TestAutoloadClasses will be
present even on Repo wikis, where they’re not necessary, but they
shouldn’t hurt either.)

Change-Id: I11eb8a14750ece91a92e74a58ba2ac657c8c2458