bridge: use @vue/cli-plugin-typescript
vue-cli-plugin-typescript seems to be a (probably outdated) copy of that
package; we should use the right one, so that it will be upgraded
together with other vue-cli packages.

ts-loader is added to package.json because we use it directly in our
Storybook config, and with just the @vue/cli-plugin-typescript change,
it was no longer found (it was buried in the plugin’s node_modules/).
I assume it only worked before thanks to npm dedupe.

Change-Id: I3641f1b58e863ee460b32d17ed9066003599b473