FP: Simulate API errors via HTTP request header
This manipulates the test system setup in such a way, that a network
error can be simulated when trying to connect to the Federated
Properties source wiki. The fake error can be triggered by sending a
custom HTTP header (e.g. via browser addon).

Such behaviour is useful for testing and validating the system in case
the source wiki's API cannot be reached.

Bug: T254999
Change-Id: I16487220523b805a0846e31cc56d7a7f93da946d