Don’t access content in EntityChange::setRevisionInfo
EntityChange::setRevisionInfo() has some code to get the entity ID from
the content of the given revision, if the entity ID is not yet
initialized. However, that code was buggy (unlike setEntityId(), it
didn’t set $this->entityId – compare the warning in getEntityId()), and
it also added an unwanted coupling between EntityChange (a Lib class)
and EntityContent (a Repo) class. It was also never used in production –
an EntityChange is always created via an EntityChangeFactory, which
takes care to call setEntityId() immediately after instantiating the
class. So let’s remove the block of code altogether and replace it with
an exception, just in case somebody tries to call setRevisionInfo() on
an EntityChange without entity ID.

Bug: T255883
Change-Id: I08ea10a8a4066572981880c99c486387cd6949fb