Reenable upstream tests for SimpleCacheWithBagOfStuff
The upstream library has a fix that should make it compatible since
2019-08, but has still not released a new version. In order to be able
to actually use the code, we copy-and-paste the file that we are
interested in and ditch the dependency on that library altogether.

This removes logging warnings and throws real exceptions again. Thus
also the tests for the logging is removed.

The original from where SimpleCacheTest.php was copied from is
https://github.com/php-cache/integration-tests/blob/e023fbe54c9821dfefbe04db0074ee8682f7f577/src/SimpleCacheTest.php

This also drops the PHPUnit 8 compatibility mechanisms from the new file
as that allows us to skip requiring the symfony/phpunit-bridge
dependency.

Bug: T237164
Change-Id: If1b68faefc830dc74422164573b1cb7c81ce4b77