Migrate Repo permissions to extension.json
On my local test wiki, this changes the order of $wgAvailableRights and
$wgGrantPermissions['editpage'], but that shouldn’t make a difference.

There is one case in which this can change the effective permissions. If
LocalSettings.php sets the permissions *after* loading the extension, as
in the following snippet, then everything is fine:

    require_once "$IP/extensions/Wikibase/repo/Wikibase.php";
    // ...
    $wgGroupPermissions['*']['property-create'] = false;

Prior to this change, Wikibase.php immediately sets $wgGroupPermissions
for that group and right to true, and then LocalSettings.php immediately
overwrites it to false, so that the effective value is false. With this
change, Wikibase.php only enqueues the registration (but does not set
$wgGroupPermissions yet), then LocalSettings.php sets the value, and
later, when the extension is processed, the permissions are updated
using wfArrayPlus2d(), which preserves existing values, so the false is
not overwritten.

However, if LocalSettings.php sets the permissions *before* loading the
extension:

    $wgGroupPermissions['*']['property-create'] = false;
    // ...
    require_once "$IP/extensions/Wikibase/repo/Wikibase.php";

Then previously, Wikibase.php would have unconditionally overwritten
$wgGroupPermissions, and the LocalSettings.php assignment would not have
had any effect; now, though, wfArrayPlus2d() will preserve the group
permissions, and anonymous users will no longer be allowed to create
properties (in this example).

Note, however, that Wikibase assigns all of the custom rights to all
users by default, and therefore this can only make a difference if
LocalSettings.php removes that right from '*' – which would previously
not have had any effect (if made before loading Wikibase.php), but with
this change now is effective. So this change in behavior doesn’t risk
making any previously forbidden behavior allowed, it only risks making
previously allowed behavior forbidden, and only if LocalSettings.php
already tried to forbid it (but in a way that was ineffective until
now).

Bug: T257441
Change-Id: I26b491f14a488c06ac64a36292e3b859b04c937d