Start using wikibase-property-notfound message again
It looks like the error message that Wikibase Client would report if it
could not find a property used to be translated, but that got lost in
change I0067e1cba4 (commit db5f210dbe), over seven years ago. It seems a
shame to throw away seven years’ worth of translations (the message had
30 translations at the time and has over 90 now), so rework the
responsible code a little bit to use the message again, making the error
translated once more: PropertyLabelNotResolvedException now extends
MessageException (a Wikibase Lib class), and the LanguageAwareRenderer
learns to handle MessageExceptions better.

Note that this changes the contents of the message a bit; specifically,
the language code is no longer shown (the old message only had a single
parameter). That seems acceptable to me, though.

The code had an old TODO to “use ExceptionLocalizer”, but that is a Repo
interface. The new code in getStatusForException() here effectively
rebuilds a small part of Repo’s ExceptionLocalizer system, specifically
the MessageExceptionLocalizer implementation. (MediaWiki core has also
had a LocalizedException class for a while – though ExceptionLocalizer
predates it – so maybe we should use that eventually. But for now Lib’s
MessageException works well enough for us.)

While we’re at it, generally modernize the code by adding some type
hints, removing PHPdoc comments that are redundant now that the type
hints are there, and declaring strict types (cf. ADR #11).

Change-Id: I9ecf52a79802fc6e62cf546c79b495d143b2636a