Migrate Magic hook handlers to hook container
The three Magic-related hook handlers look like a nice candidate for the
first use of the new hook container system (T240307) within Wikibase:
they’re relatively simple, and already handled in a dedicated class.
Previously, that class had static entry points which would create a new
instance and then call the real handler method on it; under the new
system, we instead teach extension.json how to create an instance of the
class, and then the entry points are non-static methods belonging to
three hook interfaces which the class now implements. (To reflect the
fact that the whole class is now the “handler”, rather than the three
static methods being independent “handlers”, rename the class from
…Handlers to …Handler.)

Change-Id: I3df8a3a2087b020f30388b451b492bad14edfc15