Warn for invalid keys in SimpleCacheWithBagOfStuff
This reimplements the asserting of keys for the SimpleCache, but this
time it is intended to be as permissive as PSR-16 allows.

Currently, this is only logging errors instead of throwing an exception
so that we can identify any problematic calls and investigate.

In a follow-up patch, those Exceptions will be reenabled and through
tests will be added back again.

The duplication between assertKeyIsValid and assertKeyPrefixIsValid is
kept for the moment since the functionality is just different enough
that merging them would be either a rise in complexity or a loss in
functionality and temporary enough that extracting the shared parts is
not reasonable either. It will be removed in
I14929053bac18c6f8dfb6ec9cd2a96f5dc245ce8

While integer keys are generally invalid, they are "valid" as numeric
array keys during a setMultiple call as php does not distinguish during
the array keys '0' and 0.

Also in the context of a setMultiple call, we have to already check for
the type of the key in the help method that transforms it into an
associative array as objects and arrays are valid keys in an iterable
but produce errors when used as keys in an actual array.

It finally introduces declare( strict_types = 1 ); per Wikibase ADR #11.

Bug: T246207
Bug: T245396

Change-Id: I1fb86a2c1053445b652eebcf0540893d0a2f4fb6