Set CHOKIDAR_USEPOLLING=1 for bridge test:storybook
Prior to this change the start-storybook command watches
a bunch of files and in CI it ultimately runs out of allowed
watchers.

https://github.com/paulmillr/chokidar#performance documents
some tweaks that can be made, and using polling seems to
eliminate the errors and also speed things up slightly
(around 10s performance gain).

For the CI case perhaps CHOKIDAR_INTERVAL would also give a
larger performance gain as these files will not actually be
getting changed.

This removes lots of SPAM from CI results which make it hard
to find other errors.

Bug: T255247
Change-Id: I6188e49d965c0ed5f84e19e08b7899cb75a744ce