Migrate Client SidebarHookHandler hooks to extension JSON
This class already combined three hook handlers into one; migrating it
to the hook container system therefore makes a lot of sense. (Rename the
class from …Handlers to …Handler to reflect that it’s now the whole
class that is the handler, not the individual no-longer-static methods.)

Completely remove the handler’s OtherProjectsSidebarGeneratorFactory,
which hasn’t been used since change I5ac762cb3e (commit cdefddde05).

Remove getInstance() and the private $instance, leaving only the
newFromGlobalState() factory. Caching the instance is the responsibility
of the HookContainer, not the handler class.

Remove a compatibility check (the SkinTemplateGetLanguageLink handler
being called without an $output) which hasn’t been necessary since 2014.

Remove testNewFromGlobalState() in SidebarHookHandlerTest, since the
factory is already tested in GlobalStateFactoryMethodsResourceTest.

Bug: T256260
Change-Id: I6d3372bdc618095e1992fd5112fb4a8764bee1ef