Remove $wgWBRepoDataTypes and $wgWBClientDataTypes
Instead of reading the data types in the PHP entry points, storing them
in global variables, and then accessing those global variables in the
top level factories, directly read the data type files in the factories.
Previously, other extensions *could* have modified the data types by
modifying those arrays, but they’re not supposed to do that – they
should use the WikibaseRepoDataTypes resp. WikibaseClientDataTypes hooks
instead, and as far as I can tell, all extensions also do that.

This is fairly similar to the removal of the global variables for
*entity* types in change Id27f542a01 (commit 93557f43aa). One slight
complication is the legacy $wgValueParsers global, where the parsers for
a few data types were registered under a “parser” name; we remove this
global as well, but duplicate its effects in getValueParserFactory().
(The wikibase-entityid and globecoordinate parsers are only registered
if the corresponding value types are defined; previously, in the PHP
entry point, that was guaranteed, but now some unit tests create a
WikibaseRepo with empty data type definitions.)

Bug: T256756
Change-Id: I136eed2f266a6254e407199f584810fdb8556320